feat: Implement credential preview screen
closes #190
